Move core classes and tests to core module.
