+++ build.gradle
@@ -1,25 +1,25 @@
-plugins {
-    id 'java'
+allprojects {
+    group 'org.kritiniyoga'
+    version '1.0-SNAPSHOT'
 }
 
-group 'org.kritiniyoga'
-version '1.0-SNAPSHOT'
+subprojects {
+    apply plugin: "java"
 
-sourceCompatibility = JavaVersion.VERSION_11
+    repositories {
+        mavenCentral()
+    }
 
-repositories {
-    mavenCentral()
-}
+    test {
+        useJUnitPlatform()
+    }
 
-test {
-    useJUnitPlatform()
-}
-
-dependencies {
-    compileOnly 'org.projectlombok:lombok:1.18.6'
-    annotationProcessor 'org.projectlombok:lombok:1.18.6'
-    compile 'io.vavr:vavr:0.9.3'
-    testCompile 'org.mockito:mockito-core:1.+'
-    testCompile 'org.junit.jupiter:junit-jupiter:5.4.1'
-    testCompile 'org.assertj:assertj-core:1.0.0'
+    dependencies {
+        compileOnly 'org.projectlombok:lombok:1.18.6'
+        annotationProcessor 'org.projectlombok:lombok:1.18.6'
+        compile 'io.vavr:vavr:0.9.3'
+        testCompile 'org.mockito:mockito-core:1.+'
+        testCompile 'org.junit.jupiter:junit-jupiter:5.4.1'
+        testCompile 'org.assertj:assertj-core:1.0.0'
+    }
 }
