+++ core/build.gradle
@@ -1,0 +1,2 @@
+
+
