+++ core/src/main/java/org/kritiniyoga/karmayoga/core/values/Priority.java
@@ -1,0 +1,5 @@
+package org.kritiniyoga.karmayoga.core.values;
+
+public enum Priority {
+    LOW, MEDIUM, HIGH
+}
