--- src/main/java/org/kritiniyoga/karmayoga/Allocator.java
@@ -1,7 +1,0 @@
-package org.kritiniyoga.karmayoga;
-
-import io.vavr.collection.Seq;
-
-public interface Allocator {
-    Seq<Schedule> allocate(Seq<Task> tasks, Seq<TimeSlot> slots);
-}
