--- src/main/java/org/kritiniyoga/karmayoga/Priority.java
@@ -1,5 +1,0 @@
-package org.kritiniyoga.karmayoga;
-
-public enum Priority {
-    LOW, MEDIUM, HIGH
-}
